# src/Reports/domain/responses/DRE.response.ts
export type DREResponse = {
  receitaBruta: number
  receitaLiquida: number
  custosDiretos: number
  resultadoBruto: number
  despesasOperacionais: number
  resultadoOperacional: number
  resultadosExtraordinarios: number
  resultadoLiquido: number
  year?: number
  month?: number
}